+++ demos/employees/src/main/java/is/codion/framework/demos/employees/model/EmployeeEditModel.java
@@ -61,11 +61,11 @@ public final class EmployeeEditModel extends SwingEntityEditModel {
 	private void bindEvents() {
 		//Refresh the manager ComboBoxModel when an employee is added, deleted or updated,
 		//in case a new manager got hired, fired or promoted
-		addInsertUpdateOrDeleteListener(() -> foreignKeyComboBoxModel(Employee.MANAGER_FK).refresh());
+		insertUpdateOrDeleteObserver().addListener(() -> foreignKeyComboBoxModel(Employee.MANAGER_FK).refresh());
 		//Filter the manager ComboBoxModel so that only managers from the selected department are shown,
 		//this filtering happens each time the department value is changed, either when an employee is
 		//selected or the department combo box selection changes
-		addValueListener(Employee.DEPARTMENT_FK, department -> {
+		valueObserver(Employee.DEPARTMENT_FK).addDataListener(department -> {
 			//only show managers from the same department as the selected employee and hide the currently
 			//selected employee to prevent an employee from being made her own manager
 			foreignKeyComboBoxModel(Employee.MANAGER_FK).includeCondition().set(manager ->