+++ demos/manual/src/main/java/is/codion/framework/demos/manual/notes/NotesDemo.java
@@ -121,7 +121,7 @@ public final class NotesDemo {
 		private NoteEditModel(EntityConnectionProvider connectionProvider) {
 			super(Note.TYPE, connectionProvider);
 			// Set the Note.UPDATED value before we perform an update
-			addBeforeUpdateListener(notes ->
+			beforeUpdateObserver().addDataListener(notes ->
 							notes.values().forEach(note ->
 											note.put(Note.UPDATED, LocalDateTime.now())));
 		}