+++ demos/world/src/main/java/is/codion/framework/demos/world/model/CountryEditModel.java
@@ -38,7 +38,7 @@ public final class CountryEditModel extends SwingEntityEditModel {
 	CountryEditModel(EntityConnectionProvider connectionProvider) {
 		super(Country.TYPE, connectionProvider);
 		initializeComboBoxModels(Country.CAPITAL_FK);
-		addEntityListener(country -> averageCityPopulation.set(averageCityPopulation(country)));
+		entityObserver().addDataListener(country -> averageCityPopulation.set(averageCityPopulation(country)));
 	}
 
 	@Override
@@ -46,7 +46,7 @@ public final class CountryEditModel extends SwingEntityEditModel {
 		EntityComboBoxModel comboBoxModel = super.createForeignKeyComboBoxModel(foreignKey);
 		if (foreignKey.equals(Country.CAPITAL_FK)) {
 			//only show cities for currently selected country
-			addEntityListener(country ->
+			entityObserver().addDataListener(country ->
 							comboBoxModel.includeCondition().set(city ->
 											country != null && Objects.equals(city.get(City.COUNTRY_FK), country)));
 		}