+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 - EntityEditModel.Insert, Update and Delete split into separate steps, related refactoring.
 - EntityEditModel.refreshEntity() renamed refresh().
 - EntityEditModel.referencedEntity() removed.
+- EntityEditModel, add/removeListener methods replaced with methods that return the underlying event observer.
 
 ## 0.17.34
 ### is.codion.common.core